requirements.txt
flask
scikit-learn
nltk
spacy
sentence-transformers
PyPDF2
python-docx
pdfplumber